refactor: change package structure
